third_party/blink/web_tests/external/wpt/css/css-animations/CSSAnimation-compositeOrder.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Animations are composited by their order in the animation-name property.
  assert_equals: expected "100px" but got "50px"
Harness: the test ran to completion.
